highload-systems-lab04/backend/requirements.txt
grpcio==1.71.0
grpcio-tools==1.71.0
protobuf==5.29.0
pika==1.3.2